+++ src/axom/primal/tests/primal_solid_angle.cpp
@@ -499,7 +499,6 @@ TEST(primal_solid_angle, planar_bezierpatch)
 axom::Array<primal::NURBSPatch<double, 3>> make_sphere_biquintic()
 {
   using Point3D = primal::Point<double, 3>;
-  using Vector3D = primal::Vector<double, 3>;
   using NPatch = primal::NURBSPatch<double, 3>;
 
   double rt2 = sqrt(2), rt3 = sqrt(3), rt6 = sqrt(6);
@@ -585,7 +584,6 @@ axom::Array<primal::NURBSPatch<double, 3>> make_sphere_bicubic()
 
   using Point2D = primal::Point<double, 2>;
   using Point3D = primal::Point<double, 3>;
-  using Vector3D = primal::Vector<double, 3>;
   using BCurve = primal::BezierCurve<double, 2>;
   using NPatch = primal::NURBSPatch<double, 3>;
 
@@ -666,7 +664,6 @@ TEST(primal_integral, nurbspatch_sphere)
 {
   using Point3D = primal::Point<double, 3>;
   using Vector3D = primal::Vector<double, 3>;
-  using NPatch = primal::NURBSPatch<double, 3>;
 
   int N = 12;
   // Set up an array of points to test against